+++ docs/guide/developer.rst
@@ -40,7 +40,7 @@ The first thing you need to read and understand are the base classes in the ``co
 - ``BasePolicy`` in ``policies.py`` which defines how a policy class should look like.
   It contains also all the magic for the ``.predict()`` method, to handle as many cases as possible
 
-- ``OffPolicyAlgorithm`` in ``base_class.py`` that contains the implementation of ``collect_rollouts()`` for the off-policy algorithms
+- ``OffPolicyAlgorithm`` in ``off_policy_algorithm.py`` that contains the implementation of ``collect_rollouts()`` for the off-policy algorithms
 
 
 All the environments handled internally are assumed to be ``VecEnv`` (``gym.Env`` are automatically wrapped).
