+++ stable_baselines3/a2c/a2c.py
@@ -4,7 +4,7 @@ from gym import spaces
 from typing import Type, Union, Callable, Optional, Dict, Any
 
 from stable_baselines3.common import logger
-from stable_baselines3.common.base_class import OnPolicyAlgorithm
+from stable_baselines3.common.on_policy_algorithm import OnPolicyAlgorithm
 from stable_baselines3.common.type_aliases import GymEnv, MaybeCallback
 from stable_baselines3.common.utils import explained_variance
 from stable_baselines3.common.policies import ActorCriticPolicy
