+++ stable_baselines3/common/base_class.py
@@ -1,7 +1,4 @@
 import time
-import os
-import pickle
-import warnings
 from typing import Union, Type, Optional, Dict, Any, List, Tuple, Callable
 from abc import ABC, abstractmethod
 from collections import deque
@@ -11,18 +8,16 @@ import torch as th
 import numpy as np
 
 from stable_baselines3.common import logger, utils
-from stable_baselines3.common.policies import BasePolicy, ActorCriticPolicy, get_policy_from_name
+from stable_baselines3.common.policies import BasePolicy, get_policy_from_name
 from stable_baselines3.common.utils import (set_random_seed, get_schedule_fn, update_learning_rate, get_device,
-                                            check_for_correct_spaces, safe_mean)
+                                            check_for_correct_spaces)
 from stable_baselines3.common.vec_env import DummyVecEnv, VecEnv, unwrap_vec_normalize, VecNormalize, VecTransposeImage
 from stable_baselines3.common.preprocessing import is_image_space
 from stable_baselines3.common.save_util import (recursive_getattr, recursive_setattr, save_to_zip_file,
                                                 load_from_zip_file)
-from stable_baselines3.common.type_aliases import GymEnv, RolloutReturn, MaybeCallback
+from stable_baselines3.common.type_aliases import GymEnv, MaybeCallback
 from stable_baselines3.common.callbacks import BaseCallback, CallbackList, ConvertCallback, EvalCallback
 from stable_baselines3.common.monitor import Monitor
-from stable_baselines3.common.noise import ActionNoise
-from stable_baselines3.common.buffers import ReplayBuffer, RolloutBuffer
 
 
 class BaseAlgorithm(ABC):
@@ -523,451 +518,3 @@ class BaseAlgorithm(ABC):
             params_to_save[name] = attr.state_dict()
 
         save_to_zip_file(path, data=data, params=params_to_save, tensors=tensors)
-
-
-class OffPolicyAlgorithm(BaseAlgorithm):
-    """
-    The base for Off-Policy algorithms (ex: SAC/TD3)
-
-    :param policy: Policy object
-    :param env: The environment to learn from
-                (if registered in Gym, can be str. Can be None for loading trained models)
-    :param policy_base: The base policy used by this method
-    :param learning_rate: (float or callable) learning rate for the optimizer,
-        it can be a function of the current progress remaining (from 1 to 0)
-    :param buffer_size: (int) size of the replay buffer
-    :param learning_starts: (int) how many steps of the model to collect transitions for before learning starts
-    :param batch_size: (int) Minibatch size for each gradient update
-    :param policy_kwargs: Additional arguments to be passed to the policy on creation
-    :param verbose: The verbosity level: 0 none, 1 training information, 2 debug
-    :param device: Device on which the code should run.
-        By default, it will try to use a Cuda compatible device and fallback to cpu
-        if it is not possible.
-    :param support_multi_env: Whether the algorithm supports training
-        with multiple environments (as in A2C)
-    :param create_eval_env: Whether to create a second environment that will be
-        used for evaluating the agent periodically. (Only available when passing string for the environment)
-    :param monitor_wrapper: When creating an environment, whether to wrap it
-        or not in a Monitor wrapper.
-    :param seed: Seed for the pseudo random generators
-    :param use_sde: Whether to use State Dependent Exploration (SDE)
-        instead of action noise exploration (default: False)
-    :param sde_sample_freq: Sample a new noise matrix every n steps when using gSDE
-        Default: -1 (only sample at the beginning of the rollout)
-    :param use_sde_at_warmup: (bool) Whether to use gSDE instead of uniform sampling
-        during the warm up phase (before learning starts)
-    :param sde_support: (bool) Whether the model support gSDE or not
-    """
-
-    def __init__(self,
-                 policy: Type[BasePolicy],
-                 env: Union[GymEnv, str],
-                 policy_base: Type[BasePolicy],
-                 learning_rate: Union[float, Callable],
-                 buffer_size: int = int(1e6),
-                 learning_starts: int = 100,
-                 batch_size: int = 256,
-                 policy_kwargs: Dict[str, Any] = None,
-                 tensorboard_log: Optional[str] = None,
-                 verbose: int = 0,
-                 device: Union[th.device, str] = 'auto',
-                 support_multi_env: bool = False,
-                 create_eval_env: bool = False,
-                 monitor_wrapper: bool = True,
-                 seed: Optional[int] = None,
-                 use_sde: bool = False,
-                 sde_sample_freq: int = -1,
-                 use_sde_at_warmup: bool = False,
-                 sde_support: bool = True):
-
-        super(OffPolicyAlgorithm, self).__init__(policy, env, policy_base, learning_rate,
-                                               policy_kwargs, tensorboard_log, verbose,
-                                               device, support_multi_env, create_eval_env, monitor_wrapper,
-                                               seed, use_sde, sde_sample_freq)
-        self.buffer_size = buffer_size
-        self.batch_size = batch_size
-        self.learning_starts = learning_starts
-        self.actor = None  # type: Optional[th.nn.Module]
-        self.replay_buffer = None  # type: Optional[ReplayBuffer]
-        # Update policy keyword arguments
-        if sde_support:
-            self.policy_kwargs['use_sde'] = self.use_sde
-        self.policy_kwargs['device'] = self.device
-        # For gSDE only
-        self.use_sde_at_warmup = use_sde_at_warmup
-
-    def _setup_model(self):
-        self._setup_lr_schedule()
-        self.set_random_seed(self.seed)
-        self.replay_buffer = ReplayBuffer(self.buffer_size, self.observation_space,
-                                          self.action_space, self.device)
-        self.policy = self.policy_class(self.observation_space, self.action_space,
-                                        self.lr_schedule, **self.policy_kwargs)
-        self.policy = self.policy.to(self.device)
-
-    def save_replay_buffer(self, path: str):
-        """
-        Save the replay buffer as a pickle file.
-
-        :param path: (str) Path to a log folder
-        """
-        assert self.replay_buffer is not None, "The replay buffer is not defined"
-        with open(os.path.join(path, 'replay_buffer.pkl'), 'wb') as file_handler:
-            pickle.dump(self.replay_buffer, file_handler)
-
-    def load_replay_buffer(self, path: str):
-        """
-
-        :param path: (str) Path to the pickled replay buffer.
-        """
-        with open(path, 'rb') as file_handler:
-            self.replay_buffer = pickle.load(file_handler)
-        assert isinstance(self.replay_buffer, ReplayBuffer), 'The replay buffer must inherit from ReplayBuffer class'
-
-    def collect_rollouts(self,  # noqa: C901
-                         env: VecEnv,
-                         # Type hint as string to avoid circular import
-                         callback: 'BaseCallback',
-                         n_episodes: int = 1,
-                         n_steps: int = -1,
-                         action_noise: Optional[ActionNoise] = None,
-                         learning_starts: int = 0,
-                         replay_buffer: Optional[ReplayBuffer] = None,
-                         log_interval: Optional[int] = None) -> RolloutReturn:
-        """
-        Collect rollout using the current policy (and possibly fill the replay buffer)
-
-        :param env: (VecEnv) The training environment
-        :param n_episodes: (int) Number of episodes to use to collect rollout data
-            You can also specify a ``n_steps`` instead
-        :param n_steps: (int) Number of steps to use to collect rollout data
-            You can also specify a ``n_episodes`` instead.
-        :param action_noise: (Optional[ActionNoise]) Action noise that will be used for exploration
-            Required for deterministic policy (e.g. TD3). This can also be used
-            in addition to the stochastic policy for SAC.
-        :param callback: (BaseCallback) Callback that will be called at each step
-            (and at the beginning and end of the rollout)
-        :param learning_starts: (int) Number of steps before learning for the warm-up phase.
-        :param replay_buffer: (ReplayBuffer)
-        :param log_interval: (int) Log data every ``log_interval`` episodes
-        :return: (RolloutReturn)
-        """
-        episode_rewards, total_timesteps = [], []
-        total_steps, total_episodes = 0, 0
-
-        assert isinstance(env, VecEnv), "You must pass a VecEnv"
-        assert env.num_envs == 1, "OffPolicyAlgorithm only support single environment"
-
-        if n_episodes > 0 and n_steps > 0:
-            # Note we are refering to the constructor arguments
-            # that are named `train_freq` and `n_episodes_rollout`
-            # but correspond to `n_steps` and `n_episodes` here
-            warnings.warn("You passed a positive value for `train_freq` and `n_episodes_rollout`."
-                          "Please make sure this is intended. "
-                          "The agent will collect data by stepping in the environment "
-                          "until both conditions are true: "
-                          "`number of steps in the env` >= `train_freq` and "
-                          "`number of episodes` > `n_episodes_rollout`")
-
-        if self.use_sde:
-            self.actor.reset_noise()
-
-        callback.on_rollout_start()
-        continue_training = True
-
-        while total_steps < n_steps or total_episodes < n_episodes:
-            done = False
-            episode_reward, episode_timesteps = 0.0, 0
-
-            while not done:
-
-                if self.use_sde and self.sde_sample_freq > 0 and total_steps % self.sde_sample_freq == 0:
-                    # Sample a new noise matrix
-                    self.actor.reset_noise()
-
-                # Select action randomly or according to policy
-                if self.num_timesteps < learning_starts and not (self.use_sde and self.use_sde_at_warmup):
-                    # Warmup phase
-                    unscaled_action = np.array([self.action_space.sample()])
-                else:
-                    # Note: we assume that the policy uses tanh to scale the action
-                    # We use non-deterministic action in the case of SAC, for TD3, it does not matter
-                    unscaled_action, _ = self.predict(self._last_obs, deterministic=False)
-
-                # Rescale the action from [low, high] to [-1, 1]
-                if isinstance(self.action_space, gym.spaces.Box):
-                    scaled_action = self.policy.scale_action(unscaled_action)
-
-                    # Add noise to the action (improve exploration)
-                    if action_noise is not None:
-                        # NOTE: in the original implementation of TD3, the noise was applied to the unscaled action
-                        # Update(October 2019): Not anymore
-                        scaled_action = np.clip(scaled_action + action_noise(), -1, 1)
-
-                    # We store the scaled action in the buffer
-                    buffer_action = scaled_action
-                    action = self.policy.unscale_action(scaled_action)
-                else:
-                    # Discrete case, no need to normalize or clip
-                    buffer_action = unscaled_action
-                    action = buffer_action
-
-                # Rescale and perform action
-                new_obs, reward, done, infos = env.step(action)
-
-                # Only stop training if return value is False, not when it is None.
-                if callback.on_step() is False:
-                    return RolloutReturn(0.0, total_steps, total_episodes, continue_training=False)
-
-                episode_reward += reward
-
-                # Retrieve reward and episode length if using Monitor wrapper
-                self._update_info_buffer(infos, done)
-
-                # Store data in replay buffer
-                if replay_buffer is not None:
-                    # Store only the unnormalized version
-                    if self._vec_normalize_env is not None:
-                        new_obs_ = self._vec_normalize_env.get_original_obs()
-                        reward_ = self._vec_normalize_env.get_original_reward()
-                    else:
-                        # Avoid changing the original ones
-                        self._last_original_obs, new_obs_, reward_ = self._last_obs, new_obs, reward
-
-                    replay_buffer.add(self._last_original_obs, new_obs_, buffer_action, reward_, done)
-
-                self._last_obs = new_obs
-                # Save the unnormalized observation
-                if self._vec_normalize_env is not None:
-                    self._last_original_obs = new_obs_
-
-                self.num_timesteps += 1
-                episode_timesteps += 1
-                total_steps += 1
-                if 0 < n_steps <= total_steps:
-                    break
-
-            if done:
-                total_episodes += 1
-                self._episode_num += 1
-                episode_rewards.append(episode_reward)
-                total_timesteps.append(episode_timesteps)
-
-                if action_noise is not None:
-                    action_noise.reset()
-
-                # Log training infos
-                if log_interval is not None and self._episode_num % log_interval == 0:
-                    fps = int(self.num_timesteps / (time.time() - self.start_time))
-                    logger.record("time/episodes", self._episode_num, exclude="tensorboard")
-                    if len(self.ep_info_buffer) > 0 and len(self.ep_info_buffer[0]) > 0:
-                        logger.record('rollout/ep_rew_mean', safe_mean([ep_info['r'] for ep_info in self.ep_info_buffer]))
-                        logger.record('rollout/ep_len_mean', safe_mean([ep_info['l'] for ep_info in self.ep_info_buffer]))
-                    logger.record("time/fps", fps)
-                    logger.record('time/time_elapsed', int(time.time() - self.start_time), exclude="tensorboard")
-                    logger.record("time/total timesteps", self.num_timesteps, exclude="tensorboard")
-                    if self.use_sde:
-                        logger.record("train/std", (self.actor.get_std()).mean().item())
-
-                    if len(self.ep_success_buffer) > 0:
-                        logger.record('rollout/success rate', safe_mean(self.ep_success_buffer))
-                    # Pass the number of timesteps for tensorboard
-                    logger.dump(step=self.num_timesteps)
-
-        mean_reward = np.mean(episode_rewards) if total_episodes > 0 else 0.0
-
-        callback.on_rollout_end()
-
-        return RolloutReturn(mean_reward, total_steps, total_episodes, continue_training)
-
-
-class OnPolicyAlgorithm(BaseAlgorithm):
-    """
-    The base for On-Policy algorithms (ex: A2C/PPO).
-
-    :param policy: (ActorCriticPolicy or str) The policy model to use (MlpPolicy, CnnPolicy, ...)
-    :param env: (Gym environment or str) The environment to learn from (if registered in Gym, can be str)
-    :param learning_rate: (float or callable) The learning rate, it can be a function
-        of the current progress remaining (from 1 to 0)
-    :param n_steps: (int) The number of steps to run for each environment per update
-        (i.e. batch size is n_steps * n_env where n_env is number of environment copies running in parallel)
-    :param gamma: (float) Discount factor
-    :param gae_lambda: (float) Factor for trade-off of bias vs variance for Generalized Advantage Estimator.
-        Equivalent to classic advantage when set to 1.
-    :param ent_coef: (float) Entropy coefficient for the loss calculation
-    :param vf_coef: (float) Value function coefficient for the loss calculation
-    :param max_grad_norm: (float) The maximum value for the gradient clipping
-    :param use_sde: (bool) Whether to use generalized State Dependent Exploration (gSDE)
-        instead of action noise exploration (default: False)
-    :param sde_sample_freq: (int) Sample a new noise matrix every n steps when using gSDE
-        Default: -1 (only sample at the beginning of the rollout)
-    :param tensorboard_log: (str) the log location for tensorboard (if None, no logging)
-    :param create_eval_env: (bool) Whether to create a second environment that will be
-        used for evaluating the agent periodically. (Only available when passing string for the environment)
-    :param policy_kwargs: (dict) additional arguments to be passed to the policy on creation
-    :param verbose: (int) the verbosity level: 0 no output, 1 info, 2 debug
-    :param seed: (int) Seed for the pseudo random generators
-    :param device: (str or th.device) Device (cpu, cuda, ...) on which the code should be run.
-        Setting it to auto, the code will be run on the GPU if possible.
-    :param _init_setup_model: (bool) Whether or not to build the network at the creation of the instance
-    """
-
-    def __init__(self, policy: Union[str, Type[ActorCriticPolicy]],
-                 env: Union[GymEnv, str],
-                 learning_rate: Union[float, Callable],
-                 n_steps: int,
-                 gamma: float,
-                 gae_lambda: float,
-                 ent_coef: float,
-                 vf_coef: float,
-                 max_grad_norm: float,
-                 use_sde: bool,
-                 sde_sample_freq: int,
-                 tensorboard_log: Optional[str] = None,
-                 create_eval_env: bool = False,
-                 policy_kwargs: Optional[Dict[str, Any]] = None,
-                 verbose: int = 0,
-                 seed: Optional[int] = None,
-                 device: Union[th.device, str] = 'auto',
-                 _init_setup_model: bool = True):
-
-        super(OnPolicyAlgorithm, self).__init__(policy, env, ActorCriticPolicy, learning_rate, policy_kwargs=policy_kwargs,
-                                              verbose=verbose, device=device, use_sde=use_sde, sde_sample_freq=sde_sample_freq,
-                                              create_eval_env=create_eval_env, support_multi_env=True, seed=seed)
-
-        self.n_steps = n_steps
-        self.gamma = gamma
-        self.gae_lambda = gae_lambda
-        self.ent_coef = ent_coef
-        self.vf_coef = vf_coef
-        self.max_grad_norm = max_grad_norm
-        self.rollout_buffer = None
-        self.tensorboard_log = tensorboard_log
-        self.tb_writer = None
-
-        if _init_setup_model:
-            self._setup_model()
-
-    def _setup_model(self) -> None:
-        self._setup_lr_schedule()
-        self.set_random_seed(self.seed)
-
-        self.rollout_buffer = RolloutBuffer(self.n_steps, self.observation_space,
-                                            self.action_space, self.device,
-                                            gamma=self.gamma, gae_lambda=self.gae_lambda,
-                                            n_envs=self.n_envs)
-        self.policy = self.policy_class(self.observation_space, self.action_space,
-                                        self.lr_schedule, use_sde=self.use_sde, device=self.device,
-                                        **self.policy_kwargs)
-        self.policy = self.policy.to(self.device)
-
-    def collect_rollouts(self,
-                         env: VecEnv,
-                         callback: BaseCallback,
-                         rollout_buffer: RolloutBuffer,
-                         n_rollout_steps: int) -> bool:
-
-        assert self._last_obs is not None, "No previous observation was provided"
-        n_steps = 0
-        rollout_buffer.reset()
-        # Sample new weights for the state dependent exploration
-        if self.use_sde:
-            self.policy.reset_noise(env.num_envs)
-
-        callback.on_rollout_start()
-
-        while n_steps < n_rollout_steps:
-            if self.use_sde and self.sde_sample_freq > 0 and n_steps % self.sde_sample_freq == 0:
-                # Sample a new noise matrix
-                self.policy.reset_noise(env.num_envs)
-
-            with th.no_grad():
-                # Convert to pytorch tensor
-                obs_tensor = th.as_tensor(self._last_obs).to(self.device)
-                actions, values, log_probs = self.policy.forward(obs_tensor)
-            actions = actions.cpu().numpy()
-
-            # Rescale and perform action
-            clipped_actions = actions
-            # Clip the actions to avoid out of bound error
-            if isinstance(self.action_space, gym.spaces.Box):
-                clipped_actions = np.clip(actions, self.action_space.low, self.action_space.high)
-
-            new_obs, rewards, dones, infos = env.step(clipped_actions)
-
-            if callback.on_step() is False:
-                return False
-
-            self._update_info_buffer(infos)
-            n_steps += 1
-            self.num_timesteps += env.num_envs
-
-            if isinstance(self.action_space, gym.spaces.Discrete):
-                # Reshape in case of discrete action
-                actions = actions.reshape(-1, 1)
-            rollout_buffer.add(self._last_obs, actions, rewards, dones, values, log_probs)
-            self._last_obs = new_obs
-
-        rollout_buffer.compute_returns_and_advantage(values, dones=dones)
-
-        callback.on_rollout_end()
-
-        return True
-
-    def learn(self,
-              total_timesteps: int,
-              callback: MaybeCallback = None,
-              log_interval: int = 1,
-              eval_env: Optional[GymEnv] = None,
-              eval_freq: int = -1,
-              n_eval_episodes: int = 5,
-              tb_log_name: str = "OnPolicyAlgorithm",
-              eval_log_path: Optional[str] = None,
-              reset_num_timesteps: bool = True) -> 'OnPolicyAlgorithm':
-        iteration = 0
-
-        total_timesteps, callback = self._setup_learn(total_timesteps, eval_env, callback, eval_freq,
-                                                      n_eval_episodes, eval_log_path, reset_num_timesteps,
-                                                      tb_log_name)
-
-        callback.on_training_start(locals(), globals())
-
-        while self.num_timesteps < total_timesteps:
-
-            continue_training = self.collect_rollouts(self.env, callback,
-                                                      self.rollout_buffer,
-                                                      n_rollout_steps=self.n_steps)
-
-            if continue_training is False:
-                break
-
-            iteration += 1
-            self._update_current_progress_remaining(self.num_timesteps, total_timesteps)
-
-            # Display training infos
-            if log_interval is not None and iteration % log_interval == 0:
-                fps = int(self.num_timesteps / (time.time() - self.start_time))
-                logger.record("time/iterations", iteration, exclude="tensorboard")
-                if len(self.ep_info_buffer) > 0 and len(self.ep_info_buffer[0]) > 0:
-                    logger.record("rollout/ep_rew_mean",
-                                  safe_mean([ep_info["r"] for ep_info in self.ep_info_buffer]))
-                    logger.record("rollout/ep_len_mean",
-                                  safe_mean([ep_info["l"] for ep_info in self.ep_info_buffer]))
-                logger.record("time/fps", fps)
-                logger.record("time/time_elapsed", int(time.time() - self.start_time), exclude="tensorboard")
-                logger.record("time/total_timesteps", self.num_timesteps, exclude="tensorboard")
-                logger.dump(step=self.num_timesteps)
-
-            self.train()
-
-        callback.on_training_end()
-
-        return self
-
-    def get_torch_variables(self) -> Tuple[List[str], List[str]]:
-        """
-        cf base class
-        """
-        state_dicts = ["policy", "policy.optimizer"]
-
-        return state_dicts, []
